Split the command types out of Cardano.CLI.Shelley.Parsers
So they're shared equally between the run actions and the parsers.

It also shrinks the size of the modules somewhat.
