+++ cardano-cli/cardano-cli.cabal
@@ -35,6 +35,7 @@ library
                        Cardano.CLI.Ops
                        Cardano.CLI.Parsers
                        Cardano.CLI.Run
+                       Cardano.CLI.Shelley.Commands
                        Cardano.CLI.Shelley.Parsers
                        Cardano.CLI.Shelley.Run
                        Cardano.CLI.Shelley.Run.Address
